Enum auteur_controller::CreateDestinationSubCommand

pub(crate) enum CreateDestinationSubCommand {
    Rtmp {
        id: String,
        uri: String,
    },
    LocalFile {
        id: String,
        base_name: String,
        max_size_time: Option<u32>,
    },
    LocalPlayback {
        id: String,
    },
}

Create a destination

Variants

Rtmp

Create a new RTMP destination

Fields of Rtmp

id: String

Unique identifier for the destination

uri: String

RTMP URI

LocalFile

Create a new local file destination

Fields of LocalFile

id: String

Unique identifier for the destination

base_name: String

base path, extension and potentially %05d will get appended on the other end (the latter if max_size_time is set)

max_size_time: Option<u32>

If set, the destination will split up the stream in multiple files. milliseconds

LocalPlayback

Fields of LocalPlayback

id: String

Unique identifier for the destination