Enum auteur_controller::NodeSubCommand

pub(crate) enum NodeSubCommand {
    Create {
        subcmd: CreateNodeSubCommand,
    },
    Connect {
        link_id: String,
        src_id: String,
        sink_id: String,
        config: Vec<(String, Value)>,
    },
    Disconnect {
        link_id: String,
    },
    Start {
        id: String,
        cue_time: Option<DateTime<Utc>>,
        end_time: Option<DateTime<Utc>>,
    },
    Reschedule {
        id: String,
        cue_time: Option<DateTime<Utc>>,
        end_time: Option<DateTime<Utc>>,
    },
    Remove {
        id: String,
    },
    GetInfo {
        id: Option<String>,
    },
    AddControlPoint {
        id: String,
        controllee_id: String,
        property: String,
        time: DateTime<Utc>,
        value: Value,
        mode: ArgControlMode,
    },
    RemoveControlPoint {
        id: String,
        controllee_id: String,
        property: String,
    },
}

Create and connect nodes

Variants

Create

Create a new node

Fields of Create

subcmd: CreateNodeSubCommand

Connect

Connect two existing nodes

Fields of Connect

link_id: String

The id of the link

src_id: String

The id of an existing producer node

sink_id: String

The id of an existing consumer node

config: Vec<(String, Value)>

Initial configuration of the consumer slot

Disconnect

Remove an existing link

Fields of Disconnect

link_id: String

The id of the link

Start

Cue a node for playback

Fields of Start

id: String

The id of an existing node

cue_time: Option<DateTime<Utc>>

When to cue the node, None is immediate

end_time: Option<DateTime<Utc>>

When to stop the node, None is never

Reschedule

Reschedule any node

Fields of Reschedule

id: String

The id of an existing node

cue_time: Option<DateTime<Utc>>

The new cue time. If not specified, left unchanged

end_time: Option<DateTime<Utc>>

The new end time. If not specified, left unchanged

Remove

Remove an existing node

Fields of Remove

id: String

The id of the node

GetInfo

Retrieve the info of all nodes or a specific node

Fields of GetInfo

id: Option<String>

The id of the node, if not specified, all nodes

AddControlPoint

Control properties over time for a node or slot

Fields of AddControlPoint

id: String

The id of the control point

controllee_id: String

The id of the controllee

property: String

Name of the controlled property

time: DateTime<Utc>

Time of the control point

value: Value

Desired value of the controlled property

mode: ArgControlMode

How to apply the control point

RemoveControlPoint

Remove a previously set control point

Fields of RemoveControlPoint

id: String

The id of the control point

controllee_id: String

The id of the controllee

property: String

Name of the controlled property